Fix #37, Offset the UDP base port per-CPU
Avoid port conflicts when having more than one CPU within the same
mission load CI_LAB app.  Instead of binding to the port directly,
add the CPU number as an offset.  Subtract 1 so that CPU1 remains
at the same port (1234 by default) and cmdUtil is not affected.

NOTE: the "cpu number" can be overridden on the command line when
using the pc-linux PSP, thereby also permitting the listen point
to be tuned for each launch config, if necessary.